refactor: Make YAML editor reusable
The editor is moved to it's own component. This makes it usable for
other components.

The interface to talk to the EditorComponent is as follows:

In the HTML template:
```html
<!--
  Embed the editor into your component.
  Give the element an ID, for example "editor".
  Handle submitted values with the "submitted" event handler.
 -->
<app-editor #editor (submitted)="yourOwnSubmitFunction($event)>

<!-- Add "submit value" or "reset value" buttons to your component -->
<button (click)="editor.submitValue()" />
<button (click)="editor.resetValue()" />
```

In the TS file:
```ts
// Fetch the editor component as variable
@ViewChild(EditorComponent) editor: EditorComponent | undefined;

// Set the editor value (pass JavaScript objects)
this.editor.value = {}

// Implement a submitValue handler:
yourOwnSubmitFunction(value: any){
  console.log(value);
}

```
