+++ frontend/src/app/helpers/editor/editor.component.html
@@ -1,0 +1,6 @@
+<!--
+ ~ SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
+ ~ SPDX-License-Identifier: Apache-2.0
+ -->
+
+<div id="editor" class="h-[70vh] min-h-[200px]"></div>
