+++ frontend/src/app/settings/core/configuration-settings/configuration-settings.component.html
@@ -3,13 +3,13 @@
  ~ SPDX-License-Identifier: Apache-2.0
  -->
 
-<div id="editor" class="h-[70vh] min-h-[200px]"></div>
+<app-editor #editor (submitted)="submitValue($event)" />
 <div class="m-2 flex gap-2">
-  <button mat-stroked-button (click)="submitValue()" color="primary">
+  <button mat-stroked-button (click)="editor.submitValue()" color="primary">
     <span>Save</span>
     <mat-icon class="ml-2">save</mat-icon>
   </button>
-  <button mat-stroked-button (click)="resetValue()" color="warn">
+  <button mat-stroked-button (click)="editor.resetValue()" color="warn">
     <span>Reset</span> <mat-icon class="ml-2">clear</mat-icon>
   </button>
 </div>
